requirements_bot.txt
# Liquidity Sweep Bot Requirements

# Binance API
python-binance==1.0.17

# Data processing
pandas==2.0.3
numpy==1.24.3

# Technical indicators
TA-Lib==0.4.28

# Environment variables
python-dotenv==1.0.0

# Async support
asyncio==3.4.3
websockets==12.0

# Terminal colors
termcolor==2.3.0
colorama==0.4.6

# Timezone support
pytz==2023.3

